+++ crates/factorio-cli/src/commands/login.rs
@@ -1,0 +1,28 @@
+use std::io::Write;
+
+use clap::Parser;
+use eyre::Result;
+use factorio_mod_api::ModPortalClient;
+
+use crate::App;
+
+/// Log in to the mod portal API and store the obtained login token.
+#[derive(Debug, Parser)]
+pub struct LoginCommand;
+
+impl LoginCommand {
+    pub async fn execute(&self, app: &App) -> Result<()> {
+        print!("User name or email: ");
+        std::io::stdout().flush()?;
+        let mut user_name = String::new();
+        std::io::stdin().read_line(&mut user_name)?;
+        let user_name = user_name.trim();
+
+        let password = rpassword::prompt_password("Password: ")?;
+
+        let api_token = ModPortalClient::new()?.login(user_name, &password).await?;
+        std::fs::write(app.api_token_path(), serde_json::to_string(&api_token)?)?;
+
+        Ok(())
+    }
+}